feature_extractor/app/requirements.txt
argh==0.26.2
audioread==2.1.4
cycler==0.10.0
Cython==0.25.2
decorator==4.0.11
enum34==1.1.6
functools32==3.2.3.post2
futures==3.0.5
grpcio==1.0.4
grpcio-tools==1.0.4
joblib==0.10.3
librosa==0.4.3
matplotlib==2.0.0
meld3==1.0.2
numpy==1.12.0
pandas==0.19.2
pathtools==0.1.2
protobuf==3.1.0.post1
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
resampy==0.1.4
scikit-learn==0.18.1
scipy==0.18.1
server-reloader==0.1.3
six==1.10.0
subprocess32==3.2.7
supervisor==3.3.1
virtualenv==15.1.0
watchdog==0.8.2